% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendroPlot.R, R/MeanPlot.R
\name{DendroPlot}
\alias{DendroPlot}
\alias{MeanPlot}
\title{Plot dendrogram}
\usage{
DendroPlot(Cluster.object, No.Cluster = NULL, CH = FALSE, plain = FALSE, ...)

MeanPlot(
  Cluster.object,
  No.Cluster = NULL,
  CH = FALSE,
  add.sample = 0,
  trsp = 0.5,
  seed = 123,
  ...
)
}
\arguments{
\item{Cluster.object}{The object return from function LongDataCluster}

\item{No.Cluster}{User specified number of clusters}

\item{CH}{If TRUE, iuse optimal number of clusters determined by CH index to cut dendrogram. The default is to use Gapb}

\item{plain}{If TRUE, the dendrogram only, without coloring the clustering results}

\item{...}{Additional arguments for plot}

\item{add.sample}{The number of samples plot addtionally. Default is 0 means only plot mean trajectories}

\item{trsp}{Transparent coefficient when plotting samples}

\item{seed}{Seed to draw samples from the data}
}
\description{
Plot dendrogram

Plot estimated mean trajectories for each detect cluster
}
\section{Functions}{
\itemize{
\item \code{DendroPlot}: on output from LongDataCluster, yield the corresponding dendropgram

\item \code{MeanPlot}: on output from LongDataCluster, yield the corresponding mean curves
}}

\examples{
\dontrun{
output = LongDataCluster(Longdat$Dat$obs,
Longdat$Dat[,paste("y", seq(5), sep = "_")],
Longdat$Dat$id)
DendroPlot(output)
}
\dontrun{
require(doMC)
registerDoMC(cores = 6)
output = LongDataCluster(Longdat2$Dat$obs,
Longdat2$Dat[,paste("y", seq(5), sep = "_")],
Longdat2$Dat$id, parallel = T, dropout = 15, part.size = 200)
MeanPlot(output, add.sample = 5)
}
}
