% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Longdat.R
\docType{data}
\name{Longdat}
\alias{Longdat}
\title{One example longitudinal data}
\format{
200 hundred subjects with number of observations within [4,12]. The total number of observation is 1606.
It is a list of two elements: Dat is a list of id, observation times, and 5 outcomes; Label is the right label.
}
\usage{
Longdat
}
\description{
There are four clusters exists, with size 5,65,65,65. For details, please refer to the original paper.
}
\keyword{datasets}
