% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Longdat2.R
\docType{data}
\name{Longdat2}
\alias{Longdat2}
\title{One example longitudinal data with larger sample size}
\format{
2400 hundred subjects with number of observations within [4,12]. The total number of observation is 19260
It is a list of two elements: Dat is a list of id, observation times, and 5 outcomes; Label is the right label.
}
\usage{
Longdat2
}
\description{
There are four clusters exists with size 180, 1860, 180, 180. For details, please refer to the original paper.
}
\keyword{datasets}
